+++ frontend/src/components/calendar/__test__/calendar.test.js
@@ -1,15 +1,15 @@
-import React from 'react';
-import { render, cleanup } from '@testing-library/react';
-import Calendar from '../Calendar';
-import renderer from 'react-test-renderer';
-import '@testing-library/jest-dom/extend-expect';
+import React from "react";
+import { render, cleanup } from "@testing-library/react";
+import Calendar from "../Calendar";
+import renderer from "react-test-renderer";
+import "@testing-library/jest-dom/extend-expect";
 
 // Cleaning up the rendered DOM
 afterEach(cleanup);
 
 it("Renders without problem", () => {
-    const { getByTestId } = render(<Calendar />);
-    expect(getByTestId('calendar')).not.toBeNull();
+	const { getByTestId } = render(<Calendar />);
+	expect(getByTestId("calendar")).not.toBeNull();
 });
 
 // Dummy as example of SnapShots
